[Feat] 권한별 인가 기능 추가.

+++ src/main/java/com/blogitory/blog/rolemember/entity/RoleMember.java
@@ -51,8 +51,8 @@ public class RoleMember {
   @AllArgsConstructor
   @Embeddable
   public static class Pk implements Serializable {
-    private Long roleNo;
-    private Long memberNo;
+    private Integer roleNo;
+    private Integer memberNo;
   }
 
   /**
